requirements.txt
dotenv==0.9.9
greenlet==3.1.1
mysql-connector-python==9.0.0
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
six==1.17.0
sqlalchemy==2.0.45
typing-extensions==4.13.2
tzdata==2025.3
